+++ osu.Game/Rulesets/Edit/Tools/SelectTool.cs
@@ -6,7 +6,7 @@ namespace osu.Game.Rulesets.Edit.Tools
     public class SelectTool : HitObjectCompositionTool
     {
         public SelectTool()
-            : base("Select")
+            : base("选择")
         {
         }