+++ osu.Game/Screens/Multi/Lounge/Components/DrawableRoom.cs
@@ -169,7 +169,7 @@ namespace osu.Game.Screens.Multi.Lounge.Components
                                         Spacing = new Vector2(5f),
                                         Children = new Drawable[]
                                         {
-                                            new RoomName { Font = OsuFont.GetFont(size: 18) },
+                                            new RoomName { Font = OsuFont.GetFont(size: 17) },
                                             new ParticipantInfo(),
                                         },
                                     },